Fix RISC-V sim for new binutils/gdb versions
I'm not sure when or why this broke, or why adding tconfig.h fixes it, but it
appears to.  Note that if I add tconfig.h and don't do a clean build then I get
some linking errors.

Pretty much all of this came from the bfin sim port, as that looks most similar
to ours.